Applicants Profile: Clickable Links
