+++ src/assets/main.css
@@ -8,6 +8,16 @@
   transition: all 0.3s linear;
 }
 
+.fade-enter-active,
+.fade-leave-active {
+  transition: opacity 0.3s ease;
+}
+
+.fade-enter-from,
+.fade-leave-to {
+  opacity: 0;
+}
+
 .transitionItem {
   transition: all 0.3s linear;
 }
